Add DoctrineFixtures as a dependency (even for prod environment)
Since there is not any way to create a new account (no routes for that),
we need to run the "sge:database:reinitialize" composer command whenever
we deploy a new prod app.

That said, we decided to move the DoctrineFixtures dependency from dev
dependencies to prod dependencies, at least for now.